pyproject.toml
[project]
name = "deteccion-cancer"
version = "0.1.0"
description = "Simulador de apoyo al diagnóstico de cáncer de colon (API + ML + frontend)"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "imblearn>=0.0",
    "ipykernel>=7.2.0",
    "matplotlib>=3.8.0",
    "numpy>=2.3.5",
    "openpyxl>=3.1.0",
    "pandas>=2.2.0",
    "pillow",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.0",
    "shap>=0.51.0",
    "streamlit>=1.56.0",
    "streamlit-lottie>=0.0.5",
    "torch",
    "torchvision",
    "uvicorn[standard]>=0.32.0",
    "xgboost>=3.2.0",
]

[project.optional-dependencies]
dev = [
    "httpx>=0.27.0",
    "pytest>=8.0.0",
]

[tool.pytest.ini_options]
pythonpath = [".", "backend"]
testpaths = ["tests"]
